clippy: workspace-wide auto-fix pass + manual fixes for several crates
First installment of the requested clippy cleanup pass (separate from
the earlier unsafe/panic audit). Two parts:

1. `cargo clippy --fix` applied workspace-wide across default-members,
   picking up hundreds of mechanical suggestions (needless casts/clones/
   refs, match-as-matches!, redundant patterns, into_iter->iter,
   Into/From-over-Into, if-let chains, std::fs->fs_err, etc). Added
   `fs-err` as a direct dependency to cpclib-lsp, cpclib-project and
   cpclib-dap so their disallowed-methods fixes could actually apply
   (they previously depended on it only transitively).

2. Manual fixes for issues clippy's `--fix` could not apply on its own,
   fully clearing cpclib-tokens, cpclib-basic, cpclib-image,
   cpclib-imgconverter, cpclib-cpr, cpclib-locomotive and cpclib-project:
   - Two real correctness/tool-bug findings: cpclib-image's `AsicColor`
     derived `Ord` while hand-implementing a *different* `PartialOrd`
     (clippy's `derive_ord_xor_partial_ord`, a genuine soundness risk -
     `Ord`/`PartialOrd` could disagree); cpclib-cpr's byte-array-to-
     byte-string clippy suggestion had a real precedence bug in the
     tool's own fix (`*b"RIFF".into()` parses as `*(b"RIFF".into())`,
     not `(*b"RIFF").into()`) - worked around by switching to this
     file's own existing `&str`-based `RiffCode` conversion instead.
   - One clippy false positive: cpclib-asm's `env.as_ref()` is a real
     `&mut Env` -> `&Env` downgrade a generic `Into<Fname>` bound needs,
     not the no-op `useless_asref` claims - kept, with a `#[allow]` and
     a comment explaining why.
   - Type-complexity fixes via new named type aliases (cpclib-tokens'
     `ExpandedInstruction`, cpclib-basic's `OptCommaSeparatedList`/
     `OptParenthesizedParams`).
   - too_many_arguments fixed by refactor rather than suppression per
     current direction: cpclib-image's three `ImageConverter::convert*`
     methods (8 args each) now take a `ConvertParams` struct bundling
     the palette/mode/transform/sizing parameters they all share;
     ripples into cpclib-imgconverter's one call site.
   - cpclib-image's `Palette::from_iter` inherent method (confusable
     with `std::iter::FromIterator::from_iter`) is now a real
     `FromIterator` impl; the two internal call sites that used to
     turbofish an explicit item type now disambiguate via
     `<Self as FromIterator<C>>::from_iter(..)` instead, since
     `AmstradColor: From<Rgb<u8>>` means more than one impl could apply.
   - cpclib-locomotive's `#![deny(clippy::pedantic)]` needed a real
     `# Errors` doc section on `handle_locomotive_arguments`.

Full default-members `cargo build`/`cargo test --jobs 1`: 235 test
binaries, 0 failed - verified before this commit.

Remaining crates (cpclib-catart, cpclib-dap, cpclib-lsp, cpclib-bdasm,
cpclib-runner, cpclib-bndbuild and others) still have clippy warnings
left to fix by hand; follow-up commits will cover them.

+++ cpclib-asm/src/assembler/listing_output/core.rs
@@ -1006,7 +1006,7 @@ impl ListingOutput {
                     let page = match self.current_physical_address {
                         PhysicalAddress::Memory(adr) => adr.page(),
                         PhysicalAddress::Bank(adr) => adr.bank() as u8,
-                        PhysicalAddress::Cpr(adr) => adr.bloc() as u8
+                        PhysicalAddress::Cpr(adr) => adr.bloc()
                     };
 
                     // One row per *instruction*, not per line: `ld a,l : inc a`
